genai-med-chat/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
SQLAlchemy==2.0.20
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.1
redis==4.5.5
# Optional (install when you want full features)
langchain==0.0.0 # placeholder - install specific supported version when using
qdrant-client==1.7.2
transformers==4.35.0
sentence-transformers==2.2.2
torch==2.2.0
pillow==10.0.0
pytesseract==0.3.11
pdf2image==1.16.3